YUK-9687 Change to Text Message UI for Zigbee Devices
- Added error handling for digi service commands.
- Added more fields to text message sending.

+++ yukon-client/yukon-web/src/com/cannontech/web/taglib/FormatDurationTag.java
@@ -21,7 +21,7 @@ public class FormatDurationTag extends YukonTagSupport {
     private String var;
     private boolean isVarSet = false;
     
-    private int value; //milliseconds
+    private long value; //milliseconds
     private boolean isValueSet = false; 
     
     private String type;
@@ -75,7 +75,7 @@ public class FormatDurationTag extends YukonTagSupport {
         getJspContext().getOut().print(formattedDuration);
     }
     
-    public void setValue(int value) {
+    public void setValue(long value) {
         this.value = value;
         this.isValueSet = true;
     }
